+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/MessageConverter.kt
@@ -41,6 +41,10 @@ class MessageConverter {
                 msg.video = false
             }
 
+            if (mediaOption.spotlightRequired) {
+                msg.spotlight = mediaOption.spotlight
+            }
+
             return gson.toJson(msg)
         }